% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goseq-pkg.R
\docType{package}
\name{goseq-pkg}
\alias{goseq-package}
\alias{goseq-pkg}
\title{goseq}
\description{
Detects Gene Ontology and/or other user defined categories which are 
over/under represented in RNA-seq data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/federicomarini/goseq}
  \item Report bugs at \url{https://github.com/federicomarini/goseq/issues}
}

}
\author{
\strong{Maintainer}: Federico Marini \email{marinif@uni-mainz.de} (\href{https://orcid.org/0000-0003-3252-7758}{ORCID}) [contributor]

Authors:
\itemize{
  \item Matthew Young \email{my4@sanger.ac.uk}
  \item Nadia Davidson \email{nadia.davidson@mcri.edu.au}
}

}
\keyword{internal}
