% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R
\name{pp}
\alias{pp}
\title{Prints progress through a loop}
\usage{
pp(total, count, i = i)
}
\arguments{
\item{total}{total number of iterations}

\item{count}{current iteration}

\item{i}{index of the loop}
}
\value{
message indicating the progress
}
\description{
Prints progress through a loop
}
